+++ R/datasummary_functions.R
@@ -252,7 +252,7 @@ PercentMissing <- function(x) {
 #' sometimes be useful, but is generally discouraged. Unicode characters can
 #' only display a limited number of heights for bars, and the accuracy of
 #' output is highly dependent on the platform (typeface, output type, windows
-#' vs. mac, etc.). We recommend you use the `kableExtra::spec_hist` function
+#' vs. mac, etc.). We recommend you use the `tinytable::plot_tt()` function
 #' instead.
 #'
 #' @keywords internal