+++ R/modelsummary.R
@@ -15,7 +15,7 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low',
 #' and it can produce tables in HTML, LaTeX, Word, Markdown, PDF, PowerPoint,
 #' Excel, RTF, JPG, or PNG. The appearance of the tables can be customized
 #' extensively by specifying the `output` argument, and by using functions from
-#' one of the supported table customization packages: `kableExtra`, `gt`,
+#' one of the supported table customization packages: `tinytable`, `kableExtra`, `gt`,
 #' `flextable`, `huxtable`, `DT`. For more information, see the Details and Examples
 #' sections below, and the vignettes on the `modelsummary` website:
 #' https://modelsummary.com/